+++ src/app/products/services/product.service.ts
@@ -1,15 +1,17 @@
 import { Injectable } from '@angular/core';
+import { Observable, of } from 'rxjs';
 import { Product } from '../models/product';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductService {
-  getProducts(): Product[] {
-    return [
-      { name: 'Chair', price: 10, available: true },
+  getProducts(): Observable<Product[]> {
+    return of([
+      { name: 'Lamp', price: 10, available: true },
       { name: 'Sofa', price: 40, available: true },
-      { name: 'Table', price: 80, available: false },
-    ];
+      { name: 'Chair', price: 10, available: true },
+      { name: 'Table', price: 80, available: false }
+    ]);
   }
 }